[refactor] re-organize packages
