--- pkg/instructor/anthropic_chat_stream.go
@@ -1,9 +1,0 @@
-package instructor
-
-import (
-	"context"
-)
-
-func (i *InstructorAnthropic) chatStream(ctx context.Context, request interface{}, schema *Schema) (<-chan string, error) {
-	panic("unimplemented")
-}
