+++ pkg/instructor/core/provider_enum.go
@@ -1,0 +1,10 @@
+package core
+
+type Provider = string
+
+const (
+	ProviderOpenAI    Provider = "OpenAI"
+	ProviderAnthropic Provider = "Anthropic"
+	ProviderCohere    Provider = "Cohere"
+	ProviderGoogle    Provider = "Google"
+)
