--- pkg/instructor/provider_enum.go
@@ -1,10 +1,0 @@
-package instructor
-
-type Provider = string
-
-const (
-	ProviderOpenAI    Provider = "OpenAI"
-	ProviderAnthropic Provider = "Anthropic"
-	ProviderCohere    Provider = "Cohere"
-	ProviderGoogle    Provider = "Google"
-)
